refactor: break up into smaller components
