+++ frontend/packages/schema-editor/src/components/SchemaEditor.module.css
@@ -29,12 +29,3 @@
   overflow-y: auto;
 }
 
-.addMenu {
-  margin-bottom: 2rem;
-}
-
-.dividerAbove {
-  border-top: 1px solid #c9c9c9;
-  margin-top: 4px;
-  padding-top: 4px;
-}
