+++ frontend/packages/schema-editor/src/components/layout/ModelsPanel.module.css
@@ -1,0 +1,10 @@
+
+.addMenu {
+  margin-bottom: 2rem;
+}
+
+.dividerAbove {
+  border-top: 1px solid #c9c9c9;
+  margin-top: 4px;
+  padding-top: 4px;
+}
