+++ frontend/packages/schema-editor/src/components/layout/index.ts
@@ -1,0 +1,2 @@
+export { ModelsPanel } from './ModelsPanel';
+export { TypesPanel } from './TypesPanel';
