+++ frontend/packages/schema-editor/src/components/layout/layoutTypes.ts
@@ -1,0 +1,6 @@
+import type { ILanguage } from '@altinn/schema-editor/types';
+
+export type PanelProps = {
+  language: ILanguage;
+  editMode: boolean;
+};
